feat: moves the IAgentSyncProtocol Interface declaration to a separate file
